[FrameworkBundle] Create a dedicated template filename parser
